requirements.txt
flask
flask_cors
bcrypt
pandas
python-dotenv
